# pubquiz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "API documentation for Pub Quiz application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/events/{id}/board": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Leaderboard for a given event (team, points, rank)",
                "operationId": "52773f3cfec73172068098739a8bcc6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "board": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "team": {
                                                        "type": "string",
                                                        "example": "Quiz Masters"
                                                    },
                                                    "points": {
                                                        "type": "integer",
                                                        "example": 87
                                                    },
                                                    "rank": {
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No participations found for this event."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}